@@ -25,10 +25,10 @@
     "@tanstack/react-query": "^5.8.4",
     "@tanstack/react-query-devtools": "^5.8.4",
     "clsx": "^2.0.0",
-    "next": "^13.4.19",
+    "next": "^14.0.4",
     "nodemailer": "^6.9.7",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.48.2",
     "react-hot-toast": "^2.4.1",
     "react-redux": "^9.0.0",
@@ -51,7 +51,7 @@
     "cross-env": "^7.0.3",
     "eslint": "8.54.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-next": "14.0.1",
+    "eslint-config-next": "^14.0.4",
     "eslint-config-prettier": "^9.0.0",
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-prettier": "^5.0.0",
