@@ -1,20 +1,26 @@
+"use client";
+
 /* eslint-disable @typescript-eslint/no-use-before-define */
 import { ReactNode, useLayoutEffect } from "react";
-import dynamic from "next/dynamic";
+// import dynamic from "next/dynamic";
 import { Elements } from "@stripe/react-stripe-js";
 import { useSelector } from "react-redux";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
+import dynamic from "next/dynamic";
 import getStripe from "@/helpers/GetStripe";
 import { selectCart } from "@/store/cart/CartReducer";
 import { Alert } from "@/components/shared/Alert";
+import Footer from "@/components/shared/Footer";
+import SubPageHeader from "@/components/layout/SubPageHeader";
+import BackButton from "@/components/shared/BackButton";
+// import CheckoutFormLayout from "@/components/checkout";
 
-const Footer = dynamic(import("@/components/shared/Footer"), { ssr: false });
-const Meta = dynamic(import("@/components/shared/Meta"), { ssr: false });
-const CheckoutFormLayout = dynamic(import("@/components/checkout"), { ssr: true });
-const BackButton = dynamic(import("@/components/shared/BackButton"), { ssr: false });
-const SubPageHeader = dynamic(import("@/components/layout/SubPageHeader"), { ssr: false });
+// const Meta = dynamic(import("@/components/shared/Meta"), { ssr: false });
+const CheckoutFormLayout = dynamic(() => import("@/components/checkout"));
+// const BackButton = dynamic(import("@/components/shared/BackButton"), { ssr: false });
+// const SubPageHeader = dynamic(import("@/components/layout/SubPageHeader"), { ssr: true });
 
-export default function CheckoutPage() {
+export default function CheckoutPageComponent() {
   const router = useRouter();
   const { cart } = useSelector(selectCart);
 
@@ -42,7 +48,6 @@ function CheckoutPageLayout({ children }: { children: ReactNode }) {
   const router = useRouter();
   return (
     <>
-      <Meta pageTitle="Checkout" />
       <SubPageHeader />
       <div className="bg-grey pt-[1.6rem] pb-[14.1rem] md:pt-[7.9rem] [ layout-padding ]">
         <BackButton onClick={() => router.back()} />
