@@ -1,17 +1,18 @@
 import { createContext } from "react";
-import dynamic from "next/dynamic";
 import { z } from "zod";
 import { type SubmitHandler, useForm, UseFormRegister, FieldErrors } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import toast from "react-hot-toast";
+import dynamic from "next/dynamic";
 import useOnlinePayment from "./hooks/useOnlinePayment";
-
-const CheckoutForm = dynamic(import("@/components/checkout/CheckoutForm"), { ssr: false });
-const CheckoutSummary = dynamic(import("@/components/checkout/CheckoutSummary"), { ssr: false });
-const PaymentConfirmation = dynamic(import("@/components/checkout/PaymentConfirmation"), { ssr: false });
+// import CheckoutForm from "./CheckoutForm";
 
 const mailFormat = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
 const numberFormat = /^[0-9]*$/;
+
+const PaymentConfirmation = dynamic(() => import("./PaymentConfirmation"), { ssr: false });
+const CheckoutSummary = dynamic(() => import("./CheckoutSummary"), { ssr: false });
+const CheckoutForm = dynamic(() => import("./CheckoutForm"));
 
 const inputFieldSchema = z.object({
   name: z.string().min(1, "Cannot be emtpy").min(2, "Name is required"),
