@@ -1,24 +1,24 @@
 import React, { type ReactNode } from "react";
-import dynamic from "next/dynamic";
+// import dynamic from "next/dynamic";
 import Footer from "../shared/Footer";
 import SubFooter from "../shared/SubFooter";
 import BackButton from "../shared/BackButton";
+import ProductThumbnails from "../shared/ProductThumbnails";
+import SubPageHeader from "./SubPageHeader";
 
-const Meta = dynamic(import("@/components/shared/Meta"), { ssr: false });
-const ProductThumbnails = dynamic(import("@/components//shared/ProductThumbnails"), { ssr: false });
-const SubPageHeader = dynamic(import("@/components/layout/SubPageHeader"), { ssr: false });
+// const Meta = dynamic(import("@/components/shared/Meta"), { ssr: false });
+// const ProductThumbnails = dynamic(import("@/components//shared/ProductThumbnails"), { ssr: false });
+// const SubPageHeader = dynamic(import("@/components/layout/SubPageHeader"), { ssr: false });
 
 interface Props {
   children: ReactNode;
-  pageTitle: string;
   removeSubFooter: boolean;
   onClick: () => void;
 }
 
-const ProductDetailsLayout = ({ children, pageTitle, removeSubFooter, onClick }: Props) => {
+const ProductDetailsLayout = ({ children, removeSubFooter, onClick }: Props) => {
   return (
     <>
-      <Meta pageTitle={pageTitle} />
       <SubPageHeader />
 
       <div className="[ layout-padding ]">
