@@ -1,12 +1,12 @@
+"use client";
+
 import Image from "next/image";
 import { Popover, Transition } from "@headlessui/react";
 import clsx from "clsx";
-import dynamic from "next/dynamic";
 import HamburgerIcon from "@/public/shared/tablet/icon-hamburger.svg";
 import Brand from "./Brand";
 import Cart from "../Cart/Cart";
-
-const ProductThumbnails = dynamic(import("@/components//shared/ProductThumbnails"), { ssr: false });
+import ProductThumbnails from "./ProductThumbnails";
 
 const MobileNav = () => {
   return (
