@@ -1,10 +1,10 @@
+"use client";
+
 import React from "react";
 import { clsx } from "clsx";
-import dynamic from "next/dynamic";
 import Brand from "./Brand";
 import Cart from "../Cart/Cart";
-
-const NavigationLinks = dynamic(import("./NavLink"), { ssr: false });
+import NavigationLinks from "./NavLink";
 
 const Navigation = ({ removeHero }: { removeHero: boolean }) => {
   return (
